requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
backcall==0.2.0
bleach==3.2.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
dateparser==1.0.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-spreadsheet==0.0.6
googleapis-common-protos==1.52.0
httplib2==0.18.1
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.6
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.8
pandas==1.2.0
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.9.0
prompt-toolkit==3.0.10
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.4
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-binance==0.7.9
python-dateutil==2.8.1
pytz==2020.5
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
Send2Trash==1.5.0
service-identity==18.1.0
six==1.15.0
terminado==0.9.2
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
Twisted==20.3.0
txaio==20.12.1
tzlocal==2.1
ujson==4.0.1
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zope.interface==5.2.0
